minor-fix: more helpful message in `ensure_colabfold_install` (#51)
I had forgotten to set up the logger correctly when making the package
pip-installable, which produces the undesired effect of users sitting
quietly during the first call to `sample` without providing feedback
that colabfold is being installed in the background. This PR fixes that
and provides more helpful debugging info in case the colabfold install
goes wrong.

---------

Co-authored-by: Tim Hempel <thempel@zedat.fu-berlin.de>