Revert "chore: tie colabfold dir to the python executable (#104)" (#118)
Tying the colabfold directory to the python executable seems to have
been unnecessary in retrospect and caused other headaches. We initially
thought multiple python environments cannot share a patched colabfold,
but we believe this was false, because patched colabfold is a
self-contained python environment. Therefore, I suggest we revert it.

This reverts commit afdb0f80ed43d114375b3e8f3127569055f30b78.

---------

Co-authored-by: José Jiménez-Luna <hawk31@gmail.com>