Add noise to DPMsolver (#148)
This PR tries to add an SDE version of the DPMsolver. The current
approach is to hybrid it with Heun sampler, where we first perform a
forward step of adding noise to x_t to get x_{t_hat}, and then perform a
regular DPMsolver ODE step to denoise to x_{t+dt}.

In the usage side, it will still be the `dpm.yaml` while an additional
`noise` parameter can be passed. With `noise: 0.0` we get back to the
original ODE version of DPMsolver. With the current default of `noise:
0.5` it will be the SDE version.

Empirically: if you try to push the number of steps small, e.g. 30
steps, then the ODE solver is significantly better. If you want to push
for high quality samples, then using DPMsolver SDE version or Heun
sampler with more steps (100) will be better.

---------

Co-authored-by: Yu Xie <yuxie1@microsoft.com>