chore: more informative message for empty samples after filtering (#131)
We've had a few users report about the default `--filter_samples=True`
behaviour. Currently it fails on the `mdtraj.superpose` call after
samples have been filtered, which is not very user friendly.

Instead here we add an explicit assertion to check that the number of
frames in the trajectory is greater than 0 after filtering. The error
message should be more informative.

---------

Co-authored-by: Sarah Lewis <9419264+sarahnlewis@users.noreply.github.com>